Builds an object with options for plotting likert results.

Builds an object with options for plotting likert results.

likert.options(
  low.color = "#D8B365",
  high.color = "#5AB4AC",
  neutral.color = "grey90",
  neutral.color.ramp = "white",
  colors = NULL,
  plot.percent.low = TRUE,
  plot.percent.high = TRUE,
  plot.percent.neutral = TRUE,
  plot.percents = FALSE,
  text.size = 3,
  text.color = "black",
  centered = TRUE,
  include.center = TRUE,
  ordered = TRUE,
  wrap = 50,
  wrap.grouping = 50,
  legend = "Response",
  legend.position = c("bottom", "top", "left", "right", "none"),
  panel.arrange = "v",
  panel.strip.color = "#F0F0F0",
  digits = 0,
  drop0trailing = FALSE,
  zero.print = TRUE,
  ...
)

Arguments

low.color

color for low values.

high.color

color for high values.

neutral.color

color for middle values (if odd number of levels).

neutral.color.ramp

second color used when calling colorRamp with low.color and high.color to define the color palettes.

colors

vector specifying the colors to use. This must be equal to the number of likert levels.

plot.percent.low

whether to plot low percentages.

plot.percent.high

whether to plot high percentages.

plot.percent.neutral

whether to plot netural percentages.

plot.percents

whether to label each category/bar.

text.size

size of text attributes.

text.color

color of text attributes.

centered

if true, the bar plot will be centered around zero such that the lower half of levels will be negative.

include.center

if TRUE, include the center level in the plot otherwise the center will be excluded.

ordered

reorder items from high to low.

wrap

width to wrap label text for item labels

wrap.grouping

width to wrap label text for group labels.

legend

title for the legend.

legend.position

the position for the legend ("left", "right", "bottom", "top", or two-element numeric vector).

panel.arrange

how panels for grouped likert items should be arrange. Possible values are v (vertical, the default), h (horizontal), and NULL (auto fill horizontal and vertical)

panel.strip.color

the background color for panel labels.

digits

the number of significant digits to print.

drop0trailing

logical, indicating if trailing zeros, i.e., "0" after the decimal mark, should be removed

zero.print

logical, character string or NULL specifying if and how zeros should be formatted specially.

...

included for future expansion.