Plots a set of likert items.

This is an implementation of the S3 plot generic function. Based upon the type parameter this function will call either likert.bar.plot, likert.heat.plot, or likert.density.plot. See the help pages for those functions for all the available parameters to customize the aesthetics of the figure. Although those functions can be plotted directly, we recommend call the generic plot function.

# S3 method for likert.gap
plot(
  x,
  type = c("bar", "density"),
  include.histogram = FALSE,
  panel.widths = c(3, 1),
  panel.arrange = "v",
  panel.strip.color = "#F0F0F0",
  legend.position = "bottom",
  panel.background = element_rect(size = 1, color = "grey70", fill = NA),
  satisfaction.label = "Satisfaction",
  importance.label = "Importance",
  legend,
  ...
)

Arguments

x

the likert items to plot

type

the type of plot to create. Current values are bar and heat.

include.histogram

if TRUE, a histogram of count of responses is also plotted.

panel.widths

if include.histogram=TRUE, this vector of length two specifies the ratio of the left and right panels.

panel.arrange

how panels for grouped likert items should be arrange. Possible values are v (vertical, the default), h (horizontal), and NULL (auto fill horizontal and vertical)

panel.strip.color

the background color for panel labels.

legend.position

the position for the legend ("left", "right", "bottom", "top", or two-element numeric vector).

panel.background

define background of the plot. See theme.

satisfaction.label

label used for satisfaction items.

importance.label

label used for importance items.

legend

title for the legend.

...

other parameters passed passed to likert.bar.plot or likert.heat.plot.

See also