Prints a LaTeX table of the likert items.

Create a LaTeX or HTML table of the [likert()] results.

# S3 method for class 'likert'
xtable(
  x,
  caption = NULL,
  label = NULL,
  align = NULL,
  digits = NULL,
  display = NULL,
  auto = FALSE,
  include.n = TRUE,
  include.mean = TRUE,
  include.sd = TRUE,
  include.low = TRUE,
  include.neutral = (x$nlevels%%2 != 0),
  include.high = TRUE,
  include.levels = TRUE,
  include.missing = TRUE,
  center = (x$nlevels - 1)/2 + 1,
  ordered = TRUE,
  ...
)

Arguments

x

likert class object.

caption

the table caption.

label

the table label.

align

column alignments.

digits

number of digits to use for numeric columns.

display

column formats.

auto

Logical, indicating whether to apply automatic format when no value is passed to align, digits, or display (see [xtable::xtable()] for more information.

include.n

option to include n

include.mean

option to include mean

include.sd

option to include sd

include.low

option to include low

include.neutral

option to include neutral

include.high

option to include high

include.levels

option to include levels

include.missing

option to include missing levels.

center

specifies which level should be treated as the center. For example, center = 3 would use the third level as the center whereas center = 3.5 would indicate no specific level is the center but <= 3 are low levels and >= 4 are high levels (i.e. used for forced choice items or those without a neutral option). This also influences which levels are summarized in the low and high groups.

ordered

whether the results should be ordered. See [summary.likert()]

...

other parameters passed to `xtable`.

See also